Try different approach for building "Contact Us" and "TOS" links.
For some reason, the marketing_link function used previously is unable to properly obtain the root URL for the marketing site (*), so we can't use it to build marketing links for the footer.

This solution sets the root URL of the marketing site to 'https://campus.gov.il' for both stage and prod. This is fine because the stage version of the marketing site does not include the same endpoints for the different "Contact Us" and "TOS" links, and we really only need to make sure that the endpoints of the production version of the marketing site are working.

(*) This isn't just the case for lms/templates/footer.html; lms/templates/header.html is affected as well. For the end user this doesn't make a difference, though: lms/templates/header.html just needs the root URL of the marketing site; it doesn't reference specific marketing links like CONTACT and TOS. The marketing_site function (incorrectly) returns the root URL of the LMS instead of the root URL of the marketing site *but* the root URL of the LMS is set up to redirect to the root URL of the marketing site, so the user ends up on the appropriate site anyway.