Simplify previous approach.
Marketing URLs still weren't being set correctly (despite not using the marketing_link function). We don't have the budget to investigate this further at the moment, and don't expect marketing URLs to change frequently (if at all) in the future, so referencing them directly should (and will need to) be fine for now.